--- .env
@@ -1,1 +1,0 @@
-VITE_SERVER_URL=http://localhost:3001
