--- src/Lobby.css
@@ -1,13 +1,0 @@
-.list-player{
-    border: 2px solid white;
-    height: 70vh;
-    padding: 12px;
-}
-
-.card-player{
-    border: 1px solid white;
-    padding: 8px;
-    margin-bottom: 8px;
-    border-radius: 12px 0 0 12px;
-    font-size: 1.4rem;
-}
